+++ src/app/(common)/mypage/page.tsx
@@ -3,7 +3,7 @@ import { ProfileSection } from "@/app/(common)/mypage/_components/ProfileSection
 import { MyGathering } from "@/app/(common)/mypage/types";
 import axiosInstance from "@/lib/axiosInstance";
 import getQueryClient from "@/lib/getQueryClient";
-import { GatheringType, ReviewsResponse, UserType } from "@/types";
+import { UserType } from "@/types";
 import { HydrationBoundary, dehydrate } from "@tanstack/react-query";
 
 import dayjs from "dayjs";
@@ -17,68 +17,21 @@ export default async function Page() {
   await queryClient.prefetchQuery({
     queryKey: ["user"],
     queryFn: async () => {
-      const response = await axiosInstance.get<UserType>("auths/user", {});
+      const response = await axiosInstance.get<UserType>("auths/user");
       return response.data;
     },
   });
-  await Promise.all([
-    queryClient.prefetchQuery({
-      queryKey: ["user", "gatherings"],
-      queryFn: async () => {
-        const response = await axiosInstance.get<MyGathering[]>("gatherings/joined", {
-          params: {
-            limit: 100,
-          },
-        });
-        return response.data.sort((a, b) => dayjs(b.dateTime).valueOf() - dayjs(a.dateTime).valueOf());
-      },
-    }),
-
-    queryClient.prefetchQuery({
-      queryKey: ["user", "reviews", "reviewable"],
-      queryFn: async () => {
-        const response = await axiosInstance.get<MyGathering[]>("gatherings/joined", {
-          params: {
-            limit: 100,
-            completed: true,
-            reviewed: false,
-          },
-        });
-        const data = response.data.filter((gathering) => !gathering.canceledAt);
-        return data.sort((a, b) => dayjs(b.dateTime).valueOf() - dayjs(a.dateTime).valueOf());
-      },
-    }),
-    queryClient.prefetchQuery({
-      queryKey: ["user", "reviews", "written"],
-      queryFn: async () => {
-        const user = queryClient.getQueryData<UserType>(["user"]);
-        const response = await axiosInstance.get<ReviewsResponse>("reviews", {
-          params: {
-            limit: 100,
-            userId: user?.id,
-          },
-        });
-
-        return response.data.data.sort((a, b) => dayjs(b.createdAt).valueOf() - dayjs(a.createdAt).valueOf());
-      },
-    }),
-    queryClient.prefetchQuery({
-      queryKey: ["user", "gatherings", "created"],
-      queryFn: async () => {
-        const user = queryClient.getQueryData<UserType>(["user"]);
-        const response = await axiosInstance.get<GatheringType[]>(`gatherings`, {
-          params: {
-            limit: 100,
-            sortBy: "dateTime",
-            sortOrder: "desc",
-            createdBy: user?.id,
-          },
-        });
-
-        return response.data;
-      },
-    }),
-  ]);
+  await queryClient.prefetchQuery({
+    queryKey: ["user", "gatherings"],
+    queryFn: async () => {
+      const response = await axiosInstance.get<MyGathering[]>("gatherings/joined", {
+        params: {
+          limit: 100,
+        },
+      });
+      return response.data.sort((a, b) => dayjs(b.dateTime).valueOf() - dayjs(a.dateTime).valueOf());
+    },
+  });
 
   return (
     <div className="flex flex-1 flex-col gap-[30px] pt-8">
